fix 64 to 32 bit clang conversion warning in src/core/lib

+++ src/core/lib/surface/channel.h
@@ -143,7 +143,7 @@ class Channel : public RefCounted<Channel>,
   bool is_client() const { return is_client_; }
   RegisteredCall* RegisterCall(const char* method, const char* host);
 
-  int TestOnlyRegisteredCalls() {
+  size_t TestOnlyRegisteredCalls() {
     MutexLock lock(&registration_table_.mu);
     return registration_table_.map.size();
   }
